+++ src/main/Manager.ts
@@ -53,7 +53,7 @@ export default class Manager {
 
   private cfg: ConfigService = ConfigService.getInstance();
 
-  private poller = Poller.getInstance(getFlavourConfig(this.cfg));
+  private poller = Poller.getInstance();
 
   private active = false;
 
@@ -145,8 +145,8 @@ export default class Manager {
     this.recorder = new Recorder(this.mainWindow);
 
     this.poller
-      .on('wowProcessStart', () => this.onWowStarted())
-      .on('wowProcessStop', () => this.onWowStopped());
+      .on('wowStart', () => this.onWowStarted())
+      .on('wowStop', () => this.onWowStopped());
 
     this.manage();
   }
@@ -391,7 +391,6 @@ export default class Manager {
       );
     }
 
-    this.poller.reconfigureFlavour(config);
     this.poller.start();
   }