+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make the video button more concise.
 
 ### Fixed
+- Use desktopCapturer to check if WoW is running instead of tasklist.
 
 ## [3.20.2] - 2023-11-17
 ### Fixed